Create a hooks folder and move existing hook into it
